Add logger setup and file helper functions
- Added a new module `logger.py` in the `utils` package to handle logging setup.
- Created a `setup_logging` function to configure the logging level.
- Added a new module `file_helper.py` in the `utils` package to handle file operations.
- Implemented the `ensure_directories_exist` function to create directories if they don't exist.

Closes #123
