+++ ExposureClip2Image-App/app.py
@@ -1,447 +1,35 @@
-import json
-from flask import Flask, request, render_template, send_file, url_for, Response
-from werkzeug.utils import secure_filename
-import os
-import cv2
-import numpy as np
-import glob
-import time
+from flask import Flask
 from concurrent.futures import ThreadPoolExecutor
 import logging
+import os
 
-# Initialize Flask app
-app = Flask(__name__)
+def create_app():
+    app = Flask(__name__)
 
-# Configure upload, frames, and output folders
-app.config['UPLOAD_FOLDER'] = os.path.join(app.root_path, 'static/uploads')
-app.config['FRAMES_FOLDER'] = os.path.join(app.root_path, 'static/frames')
-app.config['OUTPUT_FOLDER'] = os.path.join(app.root_path, 'outputs')
-app.config['SERVER_NAME'] = '127.0.0.1:5000'
+    app.config['UPLOAD_FOLDER'] = os.path.join(app.root_path, 'static/uploads')
+    app.config['FRAMES_FOLDER'] = os.path.join(app.root_path, 'static/frames')
+    app.config['OUTPUT_FOLDER'] = os.path.join(app.root_path, 'outputs')
+    app.config['SERVER_NAME'] = '127.0.0.1:5000'
 
-# Set up logging
-logging.basicConfig(level=logging.INFO)
+    logging.basicConfig(level=logging.INFO)
 
-# Global variable to store progress
-progress = 0
+    for folder in [app.config['UPLOAD_FOLDER'], app.config['FRAMES_FOLDER'], app.config['OUTPUT_FOLDER']]:
+        if not os.path.exists(folder):
+            os.makedirs(folder)
+            print(f"Created directory: {folder}")
 
-# Create the necessary directories if they don't exist
-for folder in [app.config['UPLOAD_FOLDER'], app.config['FRAMES_FOLDER'], app.config['OUTPUT_FOLDER']]:
-    if not os.path.exists(folder):
-        os.makedirs(folder)
-        print(f"Created directory: {folder}")
+    executor = ThreadPoolExecutor(max_workers=4)
 
-# Initialize thread pool for concurrent execution
-executor = ThreadPoolExecutor(max_workers=4)
+    from services.video_service import VideoService
+    video_service = VideoService(app)
 
-# Set up logging
-logging.basicConfig(level=logging.INFO)
+    from controllers.video_controller import create_video_controller
+    video_controller = create_video_controller(video_service, executor)
 
-def clear_frames_directory():
-    """
-    Clear all files in the frames directory.
-    """
-    try:
-        files = glob.glob(os.path.join(app.config['FRAMES_FOLDER'], '*'))
-        for f in files:
-            os.remove(f)
-        logging.info(f"Cleared frames directory: {app.config['FRAMES_FOLDER']}")
-    except Exception as e:
-        logging.error(f"Error clearing frames directory: {e}")
+    app.register_blueprint(video_controller)
 
-def extract_frames(video_path, start_time, end_time, fps, root_url):
-    """
-    Extract frames from the video between start_time and end_time at the specified fps.
+    return app
 
-    Args:
-        video_path (str): Path to the video file.
-        start_time (float): Start time in seconds.
-        end_time (float): End time in seconds.
-        fps (float): Frames per second to extract.
-        root_url (str): Root URL for frame paths.
-
-    Returns:
-        list: List of URLs for the extracted frames.
-    """
-    cap = cv2.VideoCapture(video_path)
-    original_fps = cap.get(cv2.CAP_PROP_FPS)
-    start_frame = int(start_time * original_fps)
-    end_frame = int(end_time * original_fps)
-    frame_interval = max(1, int(original_fps / fps))  # Calculate frame interval
-
-    current_frame = 0
-    frame_paths = []
-
-    try:
-        while True:
-            ret, frame = cap.read()
-            if not ret:
-                break
-            if start_frame <= current_frame <= end_frame and current_frame % frame_interval == 0:
-                frame_filename = f"frame_{current_frame}.jpg"
-                frame_path = os.path.join(app.config['FRAMES_FOLDER'], frame_filename)
-                cv2.imwrite(frame_path, frame, [cv2.IMWRITE_JPEG_QUALITY, 85])  # Save frame with JPEG quality of 85
-                frame_url = f"{root_url}static/frames/{frame_filename}"
-                frame_paths.append(frame_url)
-            current_frame += 1
-    except Exception as e:
-        logging.error(f"Error extracting frames: {e}")
-    finally:
-        cap.release()
-
-    logging.info(f"Extracted frames from {video_path}")
-    return frame_paths
-
-def create_long_exposure_image(selected_frames, highlighted_frames):
-    """
-    Create a long exposure image from the selected frames, optionally highlighting specific frames.
-
-    Args:
-        selected_frames (list): List of URLs of selected frames.
-        highlighted_frames (list): List of URLs of highlighted frames.
-
-    Returns:
-        str: Path to the created long exposure image.
-
-    Raises:
-        ValueError: If no valid frames are provided for processing.
-    """
-    exposure_image = None
-    valid_frame_count = 0
-    exposure_factor = 1.5  # Exposure factor for blending frames
-
-    for frame_url in selected_frames:
-        frame_path = frame_url.replace(request.url_root, '').replace('/static/', 'static/')
-        frame_full_path = os.path.join(app.root_path, frame_path)
-        if not os.path.exists(frame_full_path):
-            logging.warning(f"Frame file does not exist: {frame_full_path}")
-            continue
-        frame = cv2.imread(frame_full_path).astype(np.float32)
-        if exposure_image is None:
-            exposure_image = np.zeros_like(frame)
-        exposure_image += frame * exposure_factor
-        valid_frame_count += 1
-
-    if exposure_image is not None and valid_frame_count > 0:
-        exposure_image /= valid_frame_count
-
-        if highlighted_frames:
-            for highlighted_frame in highlighted_frames:
-                highlighted_path = highlighted_frame.replace(request.url_root, '').replace('/static/', 'static/')
-                highlighted_full_path = os.path.join(app.root_path, highlighted_path)
-                if os.path.exists(highlighted_full_path):
-                    highlighted_img = cv2.imread(highlighted_full_path).astype(np.float32)
-                    exposure_image = highlight_object(exposure_image, highlighted_img)
-
-        exposure_image = np.clip(exposure_image, 0, 255).astype(np.uint8)
-        timestamp = int(time.time())
-        output_path = os.path.join(app.config['OUTPUT_FOLDER'], f'long_exposure_{timestamp}.jpg')
-
-        os.makedirs(os.path.dirname(output_path), exist_ok=True)
-        logging.info(f"Saving long exposure image to: {output_path}")
-        cv2.imwrite(output_path, exposure_image)
-        return output_path
-    else:
-        raise ValueError("No valid frames to process.")
-
-def highlight_object(exposure_image, highlighted_img):
-    """
-    Highlight the main object in the highlighted image within the exposure image.
-
-    Args:
-        exposure_image (np.ndarray): The long exposure image.
-        highlighted_img (np.ndarray): The highlighted frame image.
-
-    Returns:
-        np.ndarray: The updated exposure image with the highlighted object.
-    """
-    back_sub = cv2.createBackgroundSubtractorMOG2(history=1, varThreshold=50, detectShadows=False)
-    fg_mask = back_sub.apply(highlighted_img)
-    contours, _ = cv2.findContours(fg_mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-    if contours:
-        largest_contour = max(contours, key=cv2.contourArea)
-        mask = np.zeros_like(highlighted_img, dtype=np.uint8)
-        cv2.drawContours(mask, [largest_contour], -1, (255, 255, 255), thickness=cv2.FILLED)
-        mask = mask[:, :, 0] / 255.0
-
-        alpha = 0.3  # Transparency factor for highlighted object
-        for c in range(3):
-            exposure_image[:, :, c] = (1 - mask * alpha) * exposure_image[:, :, c] + mask * alpha * highlighted_img[:, :, c]
-    return np.clip(exposure_image, 0, 255)
-
-@app.route('/', methods=['GET', 'POST'])
-def upload_file():
-    """
-    Handle file upload and frame extraction.
-
-    Returns:
-        Rendered template or redirection based on the request method.
-    """
-    if request.method == 'POST':
-        clear_frames_directory()  # Clear the frames directory before processing
-        video = request.files['video']
-        
-        # Validate and parse form data
-        try:
-            start_time = float(request.form['start_time'])
-            end_time = float(request.form['end_time'])
-            fps = float(request.form['fps'])
-        except (ValueError, KeyError) as e:
-            logging.error(f"Error parsing form data: {e}")
-            return "Invalid form data.", 400
-
-        # Validate the video file
-        if not video or not video.filename.endswith(('.mp4', '.avi', '.mov', '.mkv')):
-            logging.error("Invalid video file format.")
-            return "Invalid video file format. Only .mp4, .avi, .mov, .mkv are allowed.", 400
-
-        filename = secure_filename(video.filename)
-        video_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-        video.save(video_path)
-
-        logging.info(f"Uploaded video to {video_path}")
-
-        root_url = request.url_root
-        future = executor.submit(extract_frames, video_path, start_time, end_time, fps, root_url)
-        frame_paths = future.result()
-
-        return render_template('select_frames.html', frames=frame_paths)
-
-    return render_template('upload.html')
-
-@app.route('/process_frames', methods=['POST'])
-def process_frames():
-    """
-    Process selected frames to create a long exposure image.
-    Returns:
-        The created long exposure image file or error message.
-    """
-    try:
-        global generated_preview_path
-
-        # If the preview image has already been generated, serve it directly
-        if generated_preview_path and os.path.exists(generated_preview_path):
-            logging.info(f"Serving already generated preview image: {generated_preview_path}")
-            return send_file(generated_preview_path, as_attachment=True)
-
-        selected_frames_json = request.form.get('selectedFrames')
-        highlighted_frames_json = request.form.get('highlightedFrames')
-
-        if not selected_frames_json:
-            logging.error("No frames selected for processing.")
-            return "No frames selected for processing.", 400
-
-        selected_frames = json.loads(selected_frames_json)
-        highlighted_frames = json.loads(highlighted_frames_json) if highlighted_frames_json else []
-
-        if not selected_frames:
-            logging.error("No frames selected for processing.")
-            return "No frames selected for processing.", 400
-
-        # If not already generated, generate the long exposure image
-        output_image_path = create_long_exposure_image(selected_frames, highlighted_frames)
-        return send_file(output_image_path, as_attachment=True)
-    except json.JSONDecodeError as e:
-        logging.error(f"JSON decode error: {e}")
-        return f"An error occurred during processing: {e}. Please check the server logs for more details.", 500
-    except Exception as e:
-        logging.error(f"Error: {e}")
-        return f"An error occurred during processing: {e}. Please check the server logs for more details.", 500
-
-
-
-def apply_filter(image, filter_name):
-    if filter_name == "grayscale":
-        return cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-    elif filter_name == "sepia":
-        sepia_filter = np.array([[0.272, 0.534, 0.131],
-                                 [0.349, 0.686, 0.168],
-                                 [0.393, 0.769, 0.189]])
-        return cv2.transform(image, sepia_filter)
-    elif filter_name == "invert":
-        image_8bit = np.clip(image, 0, 255).astype(np.uint8)
-        return cv2.bitwise_not(image_8bit)
-    elif filter_name == "blur":
-        return cv2.GaussianBlur(image, (15, 15), 0)
-    elif filter_name == "sharpen":
-        kernel = np.array([[0, -1, 0],
-                           [-1, 5,-1],
-                           [0, -1, 0]])
-        return cv2.filter2D(image, -1, kernel)
-    elif filter_name == "emboss":
-        kernel = np.array([[2, 0, 0],
-                           [0, -1, 0],
-                           [0, 0, -1]])
-        return cv2.filter2D(image, -1, kernel) + 128
-    elif filter_name == "edge_detection":
-        return cv2.Canny(image.astype(np.uint8), 100, 200)
-    elif filter_name == "brightness":
-        return cv2.convertScaleAbs(image, alpha=1, beta=50)  # Increase brightness
-    elif filter_name == "contrast":
-        return cv2.convertScaleAbs(image, alpha=2.0, beta=0)  # Increase contrast
-    elif filter_name == "saturation":
-        hsv_image = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
-        hsv_image[..., 1] = cv2.add(hsv_image[..., 1], 50)  # Increase saturation
-        return cv2.cvtColor(hsv_image, cv2.COLOR_HSV2BGR)
-    elif filter_name == "posterize":
-        image_8bit = np.clip(image, 0, 255).astype(np.uint8)
-        return cv2.convertScaleAbs(image_8bit, alpha=0.5, beta=0)  # Reduce color depth
-    elif filter_name == "solarize":
-        image_8bit = np.clip(image, 0, 255).astype(np.uint8)
-        return cv2.threshold(image_8bit, 128, 255, cv2.THRESH_BINARY_INV)[1] + image_8bit
-    elif filter_name == "hdr":
-        hdr = cv2.detailEnhance(image, sigma_s=12, sigma_r=0.15)
-        return cv2.convertScaleAbs(hdr)
-    elif filter_name == "sketch":
-        gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-        inv_gray_image = 255 - gray_image
-        blur_image = cv2.GaussianBlur(inv_gray_image, (21, 21), 0)
-        sketch_image = cv2.divide(gray_image, 255 - blur_image, scale=256)
-        return cv2.cvtColor(sketch_image, cv2.COLOR_GRAY2BGR)
-    else:
-        return image  # No filter
-
-
-
-    
-generated_preview_path = None 
-
-def generate_preview(selected_frames, highlighted_frames, filter_name="none"):
-    global progress, generated_preview_path
-    exposure_image = None
-    valid_frame_count = 0
-    exposure_factor = 1.5  # Exposure factor for blending frames
-
-    def convert_url_to_path(url):
-        return os.path.join(app.root_path, url.replace(request.url_root, '').replace('/static/', 'static/'))
-
-    selected_frame_paths = [convert_url_to_path(frame_url) for frame_url in selected_frames]
-    frames_to_process = selected_frame_paths  # Always process the selected frames
-
-    total_frames = len(frames_to_process)
-
-    if total_frames == 0:
-        logging.error("No frames available for processing.")
-        raise ValueError("No frames available for processing.")
-
-    try:
-        for i, frame_path in enumerate(frames_to_process):
-            if not os.path.exists(frame_path):
-                logging.error(f"Frame path does not exist: {frame_path}")
-                continue
-
-            frame = cv2.imread(frame_path)
-            if frame is None:
-                logging.error(f"Failed to read frame: {frame_path}")
-                continue
-
-            frame = frame.astype(np.float32)
-
-            if exposure_image is None:
-                exposure_image = np.zeros_like(frame)
-
-            # Blend the frame with the exposure image
-            exposure_image += frame * exposure_factor
-            valid_frame_count += 1
-
-            # Update progress
-            progress = int(((i + 1) / total_frames) * 100)
-
-        # Normalize the image to avoid overexposure
-        if valid_frame_count > 0:
-            exposure_image /= valid_frame_count
-
-        # Process highlighted frames if any
-        if highlighted_frames:
-            for highlighted_frame in highlighted_frames:
-                highlighted_path = convert_url_to_path(highlighted_frame)
-                if os.path.exists(highlighted_path):
-                    highlighted_img = cv2.imread(highlighted_path).astype(np.float32)
-                    exposure_image = highlight_object(exposure_image, highlighted_img)
-
-        # Apply the selected filter
-        exposure_image = apply_filter(exposure_image, filter_name)
-
-        # Ensure the image values are within valid range
-        if filter_name != "grayscale":  # Grayscale already returns a single channel image
-            exposure_image = np.clip(exposure_image, 0, 255).astype(np.uint8)
-
-        # Save the generated image
-        timestamp = int(time.time())
-        output_folder = app.config['OUTPUT_FOLDER']
-        output_path = os.path.join(output_folder, f'preview_{timestamp}.jpg')
-
-        os.makedirs(output_folder, exist_ok=True)
-        logging.info(f"Saving preview image to: {output_path}")
-
-        if filter_name == "grayscale":
-            cv2.imwrite(output_path, exposure_image)  # Save grayscale image directly
-        else:
-            cv2.imwrite(output_path, exposure_image)
-
-        # Check if the file was created successfully
-        if not os.path.exists(output_path):
-            raise FileNotFoundError(f"Failed to create the preview image at {output_path}")
-
-        generated_preview_path = output_path  # Store the path of the generated preview image
-        return output_path
-
-    except Exception as e:
-        logging.error(f"Error generating preview: {e}")
-        raise
-
-@app.route('/process_preview', methods=['POST'])
-def process_preview():
-    """
-    Generate a preview of the long exposure image based on selected frames.
-    Returns:
-        The generated preview image or an error message.
-    """
-    try:
-        data = request.get_json()
-        selected_frames = data.get('selectedFrames')
-        highlighted_frames = data.get('highlightedFrames', [])
-        selected_filter = data.get('filter', 'none')  # Get selected filter
-
-        if not selected_frames:
-            logging.error("No frames selected for preview.")
-            return "No frames selected for preview.", 400
-
-        # Reset progress
-        global progress
-        progress = 0
-
-        # Generate the long exposure image for preview
-        output_image_path = generate_preview(selected_frames, highlighted_frames, selected_filter)
-
-        # Check if file exists before sending it
-        if os.path.exists(output_image_path):
-            return send_file(output_image_path, mimetype='image/jpeg')
-        else:
-            logging.error(f"File not found: {output_image_path}")
-            return "Error: Preview image could not be generated.", 500
-
-    except json.JSONDecodeError as e:
-        logging.error(f"JSON decode error: {e}")
-        return f"An error occurred during preview generation: {e}.", 500
-    except Exception as e:
-        logging.error(f"Error generating preview: {e}")
-        return f"An error occurred during preview generation: {e}.", 500
-
-
-@app.route('/progress')
-def progress_stream():
-    """
-    Stream the progress percentage to the client.
-    """
-    def generate():
-        global progress
-        while progress < 100:
-            time.sleep(0.5) 
-            yield f"data:{progress}\n\n"
-        yield "data:100\n\n"
-
-    return Response(generate(), mimetype='text/event-stream')
-
-# Run the app in debug mode
 if __name__ == '__main__':
+    app = create_app()
     app.run(debug=True)
