+++ ExposureClip2Image-App/utils/logger.py
@@ -1,0 +1,4 @@
+import logging
+
+def setup_logging():
+    logging.basicConfig(level=logging.INFO)
